# src/main/resources/static/swagger.yml
openapi: 3.0.3
info:
  title: toxic-bet
  description: <h3>Documentation for the Toxic-bet API endpoints.<h3>
  contact:
    name: Guilherme Hahn
    email: guilherme.f.h@hotmail.com
  version: latest
tags:
  - name: users
    description: Operations for uses
  - name: team
    description: Operations for teams
  - name: match
    description: Operations for matches
  - name: bet
    description: Operations for bets
  - name: betting_pool
    description: Operations for beatings pool
  - name: standings
    description: Operations for standings

paths:
  /users:
    post:
      tags:
        - users
      summary: Operation to register users in the toxic-bet application.
      description: After successfully logging into the authentication server, the user receives an email to register in the application.
      operationId: postRegisterUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDTO'
      responses:
        '201':
          description: User successfully registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    get:
      tags:
        - users
      summary: Operation to retrieve a user
      description: Retrieves detailed information about a specific user.
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: false
          description: User id to find user
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
        - name: userEmail
          in: path
          required: false
          description: User email to find user
          schema:
            type: string
            format: email
            pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
            example: "usuario@exemplo.com"
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                userNotFound:
                  summary: User not found
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "User not found."
  /match:
    post:
      tags:
        - match
      summary: Operation for registering matches
      description: It is necessary to register the matches automatically. To do this, you need to consult the table of team relationships to find the IDs. A match must have home team, away team, result set to NOT_STARTED, and the start time.
      operationId: postCreateMatch
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRequestDTO'
      responses:
        '201':
          description: Match successfully registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponseDTO'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '422':
          description: Matchmaking is not allowed.
    get:
      summary: Reactive stream of matches (SSE)
      description: Returns all existing matches and new creations of Server-Sent Events.
      operationId: streamAllMatches
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stream of matches (text/event-stream)
          content:
            text/event-stream:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                missingToken:
                  summary: Missing authentication token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Authentication token is required."
                invalidToken:
                  summary: Invalid or malformed token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Invalid authentication token."
                expiredToken:
                  summary: Expired token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Authentication token has expired."
                authenticationFailed:
                  summary: Authentication server connection failed
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Unable to authenticate. Please try again later."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    patch:
      tags:
        - match
      summary: Operation to close a match when it finish
      description: This operation will close the match, when it finish. It's a service endpoint for manual termination.
      operationId: patchCloseMatch
      security:
        - bearerAuth: []
      parameters:
        - name: matchId
          in: query
          required: true
          description: ID of the match to be closed
          schema:
            type: integer
            format: int64
        - name: result
          in: query
          required: true
          description: Result of ended match
          schema:
            type: string
            example: "HOME_WIN"
      responses:
        '204':
          description: Match Close
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                missingToken:
                  summary: Missing authentication token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Authentication token is required."
                invalidToken:
                  summary: Invalid or malformed token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Invalid authentication token."
                expiredToken:
                  summary: Expired token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Authentication token has expired."
                authenticationFailed:
                  summary: Authentication server connection failed
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Unable to authenticate. Please try again later."
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                matchNotFound:
                  summary: Match not found
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "The reported match was not found."

  /match/open:
    patch:
      tags:
        - match
      summary: Operation to open match for bet
      description: This operation will open match to bet. It's a service endpoint for manual termination.
      operationId: patchOpenMatch
      security:
        - bearerAuth: [ ]
      parameters:
        - name: matchId
          in: query
          required: true
          description: ID of the match to be closed
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Match Open
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                missingToken:
                  summary: Missing authentication token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Authentication token is required."
                invalidToken:
                  summary: Invalid or malformed token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Invalid authentication token."
                expiredToken:
                  summary: Expired token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Authentication token has expired."
                authenticationFailed:
                  summary: Authentication server connection failed
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Unable to authenticate. Please try again later."
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                matchNotFound:
                  summary: Match not found
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "The reported match was not found"
  /scores:
    get:
      summary: Operation to retrieve the results of ongoing matches.
      description: This operation will bring every second the updated result of the matches that have the result = IN_PROGRESS, informing the match ID, the names of the participating teams and the score separately.
      operationId: getStreamScores
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stream of matches (text/event-stream)
          content:
            text/event-stream:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                missingToken:
                  summary: Missing authentication token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Authentication token is required."
                invalidToken:
                  summary: Invalid or malformed token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Invalid authentication token."
                expiredToken:
                  summary: Expired token
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Authentication token has expired."
                authenticationFailed:
                  summary: Authentication server connection failed
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Unable to authenticate. Please try again later."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
  /bet:
    post:
      tags:
        - bet
      summary: Operation to register users' bets.
      description: A user can place one bet per match, containing one of the result types HOME_TEAM, DRAW, or VISITING_TEAM.
      operationId: postRegisterBet
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BetRequestDTO'
      responses:
        '201':
          description: User successfully registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetResponseDTO'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                matchNotFound:
                  summary: Match not found
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "The reported match was not found."
                userNotFound:
                  summary: user not found
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "User not found, please register."
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                matchNotFound:
                  summary: Bet already placed
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "The user has already placed a bet for this match."
        '422':
          description: Betting is not allowed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                matchAlreadyStarted:
                  summary: User has already placed a bet.
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "It is not permitted to place a bet on a match that has already started."
  /bettingPool:
    post:
      tags:
        - betting_pool
      summary: Operation to register new betting Pool
      description: Users wishing to register a betting pool must provide the pool's name. The system will return an ID and a code to identify the pool.
      operationId: postBettingPool
      security:
        - bearerAuth: [ ]
      requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BettingPoolRequestDTO'
      responses:
          '201':
            description: BettingPool successfully registered.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/BettingPoolResponseDTO'
          '400':
            description: Invalid request
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponseDTO'
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponseDTO'
    get:
      tags:
        - betting_pool
      summary: Operation to find betting pool by unique code
      description: The user who creates the betting pool shares it with other users using the unique code.
      operationId: getBettingPoolByUniqueCode
      security:
        - bearerAuth: [ ]
      parameters:
        - name: bettingPoolKey
          in: path
          required: true
          description: Unique code of the betting pool
          schema:
            type: string
            example: "A23BC45DE"
      responses:
        '200':
          description: Betting pool found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BettingPoolResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '404':
          description: Betting pool not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                bettingPoolNotFound:
                  summary: BettingPool not Found
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Not found betting pool for this key"
                userNotFound:
                  summary: User not found
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "User not found, please register."
    patch:
      tags:
        - betting_pool
      summary: Operation to add a user to the betting pool.
      description: The user uses this endpoint to join the shared betting pool.
      operationId: patchAddUserToBettingPool
      security:
        - bearerAuth: [ ]
      parameters:
        - name: bettingPoolKey
          in: path
          required: true
          description: Unique code of the betting pool
          schema:
            type: string
            example: "A23BC45DE"
      responses:
        '201':
          description: User Add To BettingPool
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '404':
          description: Betting pool not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'

  /bettingPoll/getUsers/{bettingPoolKey}:
    get:
      tags:
        - betting_pool
      summary: Operation to list users of bettingPool
      description: List the users of bettingPool order by user points desc
      operationId: getUsersByBettingPool
      security:
        - bearerAuth: [ ]
      parameters:
        - name: bettingPoolKey
          in: path
          required: true
          description: Unique code of the betting pool
          schema:
            type: string
            example: "A23BC45DE"
      responses:
        '200':
          description: Betting pool found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BettingPoolUsersResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '404':
          description: Betting pool not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
              examples:
                bettingPoolNotFound:
                  summary: BettingPool not Found
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "Not found betting pool for this key"
                userNotFound:
                  summary: User not found
                  value:
                    timestamp: "2024-12-08T19:30:00Z"
                    message: "User not found, please register."

components:
  schemas:
    ErrorResponseDTO:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
          example: "2024-01-15T10:30:00Z"
        message:
          type: string
          description: Error message description
          example: "Unauthorized access"
    SuccessResponseDTO:
      type: object
      properties:
        message:
          type: string
          description: Success message description
          example: "Success created the operation"
    MatchRequestDTO:
      type: object
      required:
        - homeTeamId
        - visitingTeamId
        - matchTime
      properties:
        homeTeamId:
          type: integer
          format: int64
          description: Home team ID (Long)
          example: 1
        visitingTeamId:
          type: integer
          format: int64
          description: Visiting team ID (Long)
          example: 2
        championshipId:
          type: integer
          format: int64
          description: Championship for which the match is valid.
          example: 1
        matchTime:
          type: string
          description: Date and time of the match in the format DD/MM/YYYY HH:MM
          pattern: "^(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[0-2])/[0-9]{4} ([01][0-9]|2[0-3]):[0-5][0-9]$"
          example: "21/12/2025 18:30"
    MatchResponseDTO:
      type: object
      properties:
        matchId:
          type: integer
          format: int64
          description: Match id (Long)
          example: 1
        homeTeamName:
          type: string
          description: Name of home team
          example: "Brazil"
        homeTeamScore:
          type: integer
          description: Score of the home team
          example: 1
        visitingTeamName:
          type: string
          description: Name of visiting team
          example: "USA"
        visitingTeamScore:
          type: integer
          description: Score of the visiting team
          example: 1
        championshipName:
          type: string
          description: Championship for which the match is valid.
          example: Premier League
        matchTime:
          type: string
          description: Date em time of match DD/MM/YYYY HH:MM
          example: "21/12/2025 18:30"
        result:
          type: string
          enum: [ NOT_STARTED, HOME_WIN, VISITING_WIN, DRAW, IN_PROGRESS, OPEN_FOR_BETTING, ENDED, CLOSE_FOR_BETTING ]
          description: Match result
          example: "NOT_STARTED"
        homeTeamOdds:
          type: number
          format: double
          description: Odds for home team
          example: "2.5"
        visitingTeamOdds:
          type: number
          format: double
          description: Odds for home team
          example: "2.5"
        drawTeamOdds:
          type: number
          format: double
          description: Odds for home team
          example: "2.5"
    EndedMatchRequestDTO:
      type: object
      required:
        - result
      properties:
        result:
          type: string
          enum: [ ENDED ]
          description: Match result
          example: "ENDED"
    UserRequestDTO:
      type: object
      required:
        - name
        - email
      properties:
        email:
          type: string
          format: email
          description: User email address to be registered
          example: usuario@exemplo.com
        name:
          type: string
          description: User name to be registered
          example: John Coffee
    UserResponseDTO:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        userName:
          type: string
          example: "John Doe"
        userEmail:
          type: string
          format: email
          example: "johnDoe@email.com"
    UpdateScoreRequestDTO:
      type: object
      required:
        - homeTeamScore
        - visitingTeamScore
      properties:
        homeTeamScore:
          type: integer
          minimum: 0
          description: Score of the home team
          example: 2
        visitingTeamScore:
          type: integer
          minimum: 0
          description: Score of the visiting team
          example: 1
    BetRequestDTO:
      type: object
      required:
        - matchId
        - result
        - odds
      properties:
        matchId:
          type: integer
          format: int64
          description: Match ID for which the bet is being placed.
          example: 2
        result:
          type: string
          enum: [ HOME_WIN, VISITING_WIN, DRAW ]
          description: Match result
          example: "HOME_WIN"
        odds:
          type: number
          format: double
          description: The odds value that will be used to calculate the user's score.
          example: 1.75
    BetResponseDTO:
      type: object
      properties:
        betId:
          type: integer
          format: int64
          description: ID of the bet placed by the user.
          example: 1
        matchId:
          type: integer
          format: int64
          description: Match ID where the user placed the bet.
          example: 1
        result:
          type: string
          description: The result the user bet on.
          example: "HOME_WIN"
        odds:
          type: number
          format: double
          description: Odds the user will receive if they win the bet.
          example: 2.45
    BettingPoolRequestDTO:
      type: object
      properties:
        bettingPoolName:
          type: string
          description: Name of bettingPool
          example: "Old guys toxic bet"
    BettingPoolUsersResponseDTO:
      type: object
      properties:
        bettingPoolId:
          type: integer
          format: int64
          description: Betting Pool ID
          example: 1
        bettingPoolKey:
          type: string
          description: Unique Key of BettingPoll
          example: "A23BC45DE"
        users:
          type: object
          description: Map userName
          additionalProperties:
            type: number
            format: double
          example:
            "John Coffee": 12.5
            "Jane Doe": 7.0
    BettingPoolResponseDTO:
      type: object
      properties:
        idBettingPool:
          type: integer
          format: int64
          description: ID of the betting poll
          example: 1
        bettingPoolName:
          type: string
          description: Name of betting pool
          example: "Old guys toxic bet"
        bettingPoolKey:
          type: string
          description: Unique code of betting pool
          example: "A23BC45DE"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token generated by the authentication server to be validated in the application.
security:
  - bearerAuth: []